+++ R/app_ui.R
@@ -18,10 +18,10 @@ app_ui <- function(request) {
       
       # Title header
       header = shinydashboard::dashboardHeader(
-        title = "GlycoDash v1.9.2",
+        title = paste0("GlycoDash v", utils::packageVersion("GlycoDash")),
         tags$li(
           ManualButton("download_manual"), 
-          title = "Download latest version of user guide",
+          title = "Download user guide",
           class = "dropdown"
         ),
         tags$li(