+++ R/mod_normalization.R
@@ -146,6 +146,7 @@ mod_normalization_ui <- function(id){
   )
 }
     
+
 #' normalization Server Functions
 #'
 #' @noRd 
@@ -154,17 +155,13 @@ mod_normalization_server <- function(id, results_analyte_curation, merged_metada
     ns <- session$ns
     
     analyte_curated_data <- reactive({
-      # req() considers an empty dataframe Truthy, and because of the way that
-      # results_analyte_curation$analyte_curated_data() is created in
-      # mod_analyte_curation.R there is a moment that it is an empty dataframe.
-      # Solution -> wait until results_analyte_curation$analyte_curated_data()
-      # is not empty:
       req(!rlang::is_empty(results_analyte_curation$analyte_curated_data()))
       # Check if charge states should be treated separately
-      if (input$separate_charges == TRUE) {
+      if (input$separate_charges) {
         results_analyte_curation$analyte_curated_data() %>% 
           tidyr::unite("analyte", analyte:charge, sep = "_", remove = FALSE)
-      } else {
+      } 
+      else {
         results_analyte_curation$analyte_curated_data()
       }
     }) 
@@ -182,19 +179,25 @@ mod_normalization_server <- function(id, results_analyte_curation, merged_metada
     # Normalized data is in long format
     normalized_data <- reactive({
       req(total_intensities_glycans())
-      data <- normalize_data(total_intensities = total_intensities_glycans()) %>% 
-        # Sort by glycan composition
-        tidyr::separate(analyte, into = c("cluster", "analyte"),
-                        sep = "1", extra = "merge") %>% 
-        sort_glycans(.) %>% 
-        dplyr::group_by(cluster) %>% 
-        dplyr::arrange(cluster, analyte) %>% 
-        dplyr::ungroup() %>% 
-        dplyr::mutate(analyte = paste0(cluster, "1", analyte))
+      if (nrow(total_intensities_glycans()) == 0) {
+        # Zero analytes passed
+        data <- normalize_data(total_intensities = total_intensities_glycans())
+      }
+      else {
+        data <- normalize_data(total_intensities = total_intensities_glycans()) %>% 
+          # Sort by glycan composition
+          tidyr::separate(analyte, into = c("cluster", "analyte"),
+                          sep = "1", extra = "merge") %>% 
+          sort_glycans(.) %>% 
+          dplyr::group_by(cluster) %>% 
+          dplyr::arrange(cluster, analyte) %>% 
+          dplyr::ungroup() %>% 
+          dplyr::mutate(analyte = paste0(cluster, "1", analyte))
         
-      # Check if metadata exists
-      if (is_truthy(merged_metadata())) {
-        data <- dplyr::left_join(data, merged_metadata(), by = "sample_id")
+        # Check if metadata exists
+        if (is_truthy(merged_metadata())) {
+          data <- dplyr::left_join(data, merged_metadata(), by = "sample_id")
+        }
       }
       
       return(data)
@@ -276,24 +279,28 @@ mod_normalization_server <- function(id, results_analyte_curation, merged_metada
         
         # Create tabs and plots
         purrr::imap(cluster_names, function(cluster, i) {
-          
-          plot <- sample_heatmap(
-            normalized_data = normalized_data(),
-            cluster_name = cluster,
-            exclude_sample_types = input$exclude_sample_types,
-            group_facet = dplyr::case_when(
-              .default = "", 
-              ( # Only when switch is on, and curation was done per biological group.
-                # Second check is required, because switch is on by default.
-                input$facet_per_group == TRUE &
-                results_analyte_curation$curation_method() == "Per biological group"
-              ) ~ results_analyte_curation$biogroups_colname()
-            ),
-            color_low = input$color_low,
-            color_mid = input$color_mid,
-            color_high = input$color_high,
-            color_na = input$color_na
-          )
+          if (nrow(normalized_data()) == 0) {
+            plot <- ggplot2::ggplot()
+          }
+          else {
+            plot <- sample_heatmap(
+              normalized_data = normalized_data(),
+              cluster_name = cluster,
+              exclude_sample_types = input$exclude_sample_types,
+              group_facet = dplyr::case_when(
+                .default = "", 
+                ( # Only when switch is on, and curation was done per biological group.
+                  # Second check is required, because switch is on by default.
+                  input$facet_per_group &
+                    results_analyte_curation$curate_per_group()
+                ) ~ results_analyte_curation$biogroups_colname()
+              ),
+              color_low = input$color_low,
+              color_mid = input$color_mid,
+              color_high = input$color_high,
+              color_na = input$color_na
+            )
+          }
           
           # Store plot in list, give it name of the cluster.
           # Need isolate() to prevent infinite loop
@@ -303,7 +310,9 @@ mod_normalization_server <- function(id, results_analyte_curation, merged_metada
           })
           
           # Show plot in UI
-          output[[cluster]] <- plotly::renderPlotly(plotly::ggplotly(plot, tooltip = "text"))
+          output[[cluster]] <- plotly::renderPlotly(
+            plotly::ggplotly(plot, tooltip = "text")
+          )
           appendTab(
             inputId = "tabs",
             select = TRUE,
@@ -316,29 +325,36 @@ mod_normalization_server <- function(id, results_analyte_curation, merged_metada
         
         #### CLUSTER ON Y-AXIS #### 
       } else if (input$heatmap_yaxis == "Glycosylation site") {
-
+      
         # Make the plot
-        plot <- cluster_heatmap(
-          normalized_data = normalized_data(),
-          exclude_sample_types = input$exclude_sample_types,
-          group_facet = dplyr::case_when(
-            .default = "",
-            (
-              input$facet_per_group == TRUE &
-              results_analyte_curation$curation_method() == "Per biological group"
-            ) ~ results_analyte_curation$biogroups_colname()
-          ),
-          color_low = input$color_low,
-          color_mid = input$color_mid,
-          color_high = input$color_high,
-          color_na = input$color_na
-        )
+        if (nrow(normalized_data()) == 0) {
+          plot <- ggplot2::ggplot()
+        } 
+        else {
+          plot <- cluster_heatmap(
+            normalized_data = normalized_data(),
+            exclude_sample_types = input$exclude_sample_types,
+            group_facet = dplyr::case_when(
+              .default = "",
+              (
+                input$facet_per_group &
+                  results_analyte_curation$curate_per_group()
+              ) ~ results_analyte_curation$biogroups_colname()
+            ),
+            color_low = input$color_low,
+            color_mid = input$color_mid,
+            color_high = input$color_high,
+            color_na = input$color_na
+          )
+        }
         
         # Store plot in list
         isolate(r$heatmaps <- list(plot))
         
         # Show plot in UI and store in list
-        output$clusters_plot <- plotly::renderPlotly(plotly::ggplotly(plot, tooltip = "text"))
+        output$clusters_plot <- plotly::renderPlotly(
+          plotly::ggplotly(plot, tooltip = "text")
+        )
       }
       
     }) %>% 
@@ -353,8 +369,8 @@ mod_normalization_server <- function(id, results_analyte_curation, merged_metada
     
     # Excluded sample types, to pass on to report
     heatmaps_excluded_sample_types <- reactive({
-      if (results_analyte_curation$curation_method() == "Per biological group") {
-        if (input$facet_per_group == TRUE) {
+      if (results_analyte_curation$curate_per_group()) {
+        if (input$facet_per_group) {
           c("")
         } else if (length(input$exclude_sample_types) == 0) {
           c("None")
@@ -385,19 +401,17 @@ mod_normalization_server <- function(id, results_analyte_curation, merged_metada
     observe({
       shinyjs::toggleState("download", is_truthy(normalized_data_wide()))
       
-      if (results_analyte_curation$curation_method() == "Per biological group" &
-          results_analyte_curation$biogroups_colname() != "") {
-        shinyjs::show("facet_per_group")
-      } else {
+      if (!results_analyte_curation$curate_per_group()) {
         shinyjs::hide("facet_per_group")
-      }
-      
-      
-      if (results_analyte_curation$curation_method() == "Per biological group" &
-          input$facet_per_group == TRUE & results_analyte_curation$biogroups_colname() != "") {
-        shinyjs::hide("exclude_sample_types")
-      } else {
         shinyjs::show("exclude_sample_types")
+        
+      } else {
+        shinyjs::show("facet_per_group")
+        if (input$facet_per_group) {
+          shinyjs::hide("exclude_sample_types")
+        } else {
+          shinyjs::show("exclude_sample_types")
+        }
       }
       
       if (input$heatmap_yaxis == "Sample") {
@@ -423,11 +437,13 @@ mod_normalization_server <- function(id, results_analyte_curation, merged_metada
       },
       content = function(file) {
         if (grepl("R object", input$download_format)) {
-          save(normalized_data_wide(), file = file)
-        } else if (is_truthy(notes())) {
+          saveRDS(normalized_data_wide(), file = file)
+        } 
+        else if (is_truthy(notes())) {
           data_list <- list("Data" = normalized_data_wide(), "Notes" = notes())
           writexl::write_xlsx(data_list, path = file)
-        } else{
+        } 
+        else{
           writexl::write_xlsx(normalized_data_wide(), path = file)
         }
       }